LayoutTests/http/tests/security/xssAuditor/base-href-scheme-relative-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-head-base-href.pl?q=%3Cbase%20href='//127.0.0.1:8000/security/xssAuditor/resources/base-href/'%3E' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
ALERT: This is a safe script.
